Revert "Resubmit rL345008 "Split MachinePipeliner code into header and cpp files""
This reverts commit rL350493
issues related to modules  still appear in http://green.lab.llvm.org/green/job/lldb-cmake

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@350497 91177308-0d34-0410-b5e6-96231b3b80d8
